+++ core/src/executive/executive.rs
@@ -1442,7 +1442,7 @@ impl<'a> Executive<'a> {
                 Executed::not_enough_balance_fee_charged(
                     tx,
                     &actual_gas_cost,
-                    &self.env.gas_used,
+                    &self.env.gas_charged,
                 ),
             ));
         } else {
@@ -1479,7 +1479,7 @@ impl<'a> Executive<'a> {
                         ExecutionError::ContractAddressConflict,
                         Executed::execution_error_fully_charged(
                             tx,
-                            &self.env.gas_used,
+                            &self.env.gas_charged,
                         ),
                     ));
                 }
@@ -1540,14 +1540,21 @@ impl<'a> Executive<'a> {
             None
         };
 
-        Ok(self.finalize(tx, substate, result, output, refund_receiver)?)
+        Ok(self.finalize(
+            tx,
+            substate,
+            result,
+            output,
+            refund_receiver,
+            storage_sponsored,
+        )?)
     }
 
     /// Finalizes the transaction (does refunds and suicides).
     fn finalize(
         &mut self, tx: &SignedTransaction, mut substate: Substate,
         result: vm::Result<FinalizationResult>, output: Bytes,
-        refund_receiver: Option<Address>,
+        refund_receiver: Option<Address>, storage_sponsor_paid: bool,
     ) -> DbResult<ExecutionOutcome>
     {
         let gas_left = match result {
@@ -1560,20 +1567,16 @@ impl<'a> Executive<'a> {
         // gas_left should be smaller than 1/4 of gas_limit, otherwise
         // 3/4 of gas_limit is charged.
         let charge_all = (gas_left + gas_left + gas_left) >= gas_used;
-        let (cumulative_gas_used, fees_value, refund_value) = if charge_all {
+        let (gas_charged, fees_value, refund_value) = if charge_all {
             let gas_refunded = tx.gas >> 2;
             let gas_charged = tx.gas - gas_refunded;
             (
-                self.env.gas_used + gas_charged,
+                gas_charged,
                 gas_charged * tx.gas_price,
                 gas_refunded * tx.gas_price,
             )
         } else {
-            (
-                self.env.gas_used + gas_used,
-                gas_used * tx.gas_price,
-                gas_left * tx.gas_price,
-            )
+            (gas_used, gas_used * tx.gas_price, gas_left * tx.gas_price)
         };
 
         if let Some(r) = refund_receiver {
@@ -1616,7 +1619,10 @@ impl<'a> Executive<'a> {
             Err(vm::Error::StateDbError(e)) => bail!(e),
             Err(exception) => Ok(ExecutionOutcome::ExecutionErrorBumpNonce(
                 ExecutionError::VmError(exception),
-                Executed::execution_error_fully_charged(tx, &self.env.gas_used),
+                Executed::execution_error_fully_charged(
+                    tx,
+                    &self.env.gas_charged,
+                ),
             )),
             Ok(r) => {
                 let mut storage_collateralized = Vec::new();
@@ -1661,10 +1667,12 @@ impl<'a> Executive<'a> {
                 let executed = Executed {
                     gas: tx.gas,
                     gas_used,
+                    gas_charged,
                     fee: fees_value,
-                    cumulative_gas_used,
+                    gas_sponsor_paid: refund_receiver.is_some(),
                     logs: substate.logs,
                     contracts_created: substate.contracts_created,
+                    storage_sponsor_paid,
                     storage_collateralized,
                     storage_released,
                     output,