+++ primitives/src/receipt.rs
@@ -21,33 +21,46 @@ pub struct StorageChange {
 /// Information describing execution of a transaction.
 #[derive(Debug, Clone, PartialEq, Eq, RlpDecodable, RlpEncodable)]
 pub struct Receipt {
-    /// The total gas used in the block following execution of the transaction.
+    /// The total gas charged in the block following execution of the
+    /// transaction.
+    // TODO: change the name to gas_charged.
     pub gas_used: U256,
+    /// The total gas cost value in the block following execution of the
+    /// transaction.
+    pub gas_fee: U256,
+    /// The designated account to bear the gas fee, if any.
+    pub gas_sponsor_paid: bool,
     /// The OR-wide combination of all logs' blooms for this transaction.
     pub log_bloom: Bloom,
     /// The logs stemming from this transaction.
     pub logs: Vec<LogEntry>,
     /// Transaction outcome.
     pub outcome_status: u8,
+    /// The designated account to bear the storage fee, if any.
+    pub storage_sponsor_paid: bool,
     pub storage_collateralized: Vec<StorageChange>,
     pub storage_released: Vec<StorageChange>,
 }
 
 impl Receipt {
     pub fn new(
-        outcome: u8, gas_used: U256, logs: Vec<LogEntry>,
+        outcome: u8, gas_charged: U256, gas_fee: U256, gas_sponsor_paid: bool,
+        logs: Vec<LogEntry>, storage_sponsor_paid: bool,
         storage_collateralized: Vec<StorageChange>,
         storage_released: Vec<StorageChange>,
     ) -> Self
     {
         Self {
-            gas_used,
+            gas_used: gas_charged,
+            gas_fee,
+            gas_sponsor_paid,
             log_bloom: logs.iter().fold(Bloom::default(), |mut b, l| {
                 b.accrue_bloom(&l.bloom());
                 b
             }),
             logs,
             outcome_status: outcome,
+            storage_sponsor_paid,
             storage_collateralized,
             storage_released,
         }